refactoring : redis global 이동
