+++ src/main/java/poomasi/global/config/redis/config/RedisConfig.java
@@ -1,4 +1,4 @@
-package poomasi.domain.auth.token.redis.config;
+package poomasi.global.config.redis.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
