+++ src/main/java/poomasi/global/config/redis/error/RedisConnectionException.java
@@ -1,4 +1,4 @@
-package poomasi.domain.auth.token.redis.error;
+package poomasi.global.config.redis.error;
 
 public class RedisConnectionException extends RuntimeException {
     public RedisConnectionException(String message) {
