+++ src/main/java/poomasi/global/config/redis/error/RedisOperationException.java
@@ -1,4 +1,4 @@
-package poomasi.domain.auth.token.redis.error;
+package poomasi.global.config.redis.error;
 
 public class RedisOperationException extends RuntimeException {
     public RedisOperationException(String message) {
